#7 : ENHC : Worker thread Implementation from ThulawaProcessor
Updated the classes ThulawaScheduler, QueueManager, ThreadPoolRegistry as Singleton instances.

However do need to handle Circular dependencies arising from them. Anyway this initial implementation works as Stream processing Engine by processing the arriving records.

The Scheduling logic needs work, and will be attended in the coming weeks.

A code explanation with the architecture will be provided in the Notion. Since the implementation is satisfactory will be closing the issue.

The Scheduling logic will be added after creating a separate issue. Perhaps, will be updating the Notion with Streams Code example as well.

This issue will be closed and will be merged with the master branch!