#15 : ENHC : Refactoring Thulawa Metrics with proper design pattern
With these updates, A ThulawaMetrics object will be created and will be saved as a Stream config and then it can be accessed using the THULAWA_METRICS_CONFIG from the processor context which is present in the ThulawaProcessor.

Also initial recorder class related to issue - #22, which is to manage JVM metrics is also included in this commit.