#32 : ENHC : Implement the Filtering based on the Key
With the following implementation, Records are distributed among Queues based on the priority of the Key. For each high priority key a separate Queue is created and all other keys are sent to the low priority queue.

This works fine with the multi-threaded implementation, just have to manage the thread queue size, and of threads